widget: move `client` into own module and split up
